refactor(settings): overhaul About screen and Settings labels, fix typos and GCC 15 compat
- Rewrite About Device screen to show actual hardware info (Display, Chip,
  RAM, Storage, Battery) in bracketed values
- Shorten Settings labels: Animation, Reset Defaults, Storage
- Show battery type as [LiPo]/[CR2032] instead of ON/OFF
- Bracket OFF values in backlight and standby timeout
- Rename vfs_stat_t.avaliable -> available (10 usages)
- Rename NFCID1_*_LENGHT -> NFCID1_*_LENGTH (2 defines, 5 usages)
- Fix typos: avaliable_languages, CMASS-DAP, programer, paramter, etc.
- Add (uintptr_t) casts in pointer-to-int conversions for GCC 15
- Add forward declaration in chameleon_scene_menu_card_advanced.c
- Remove static from enum in amiidb_scene_amiibo_detail_menu.c
- Suppress new GCC 15 warnings in Makefile, remove -u _printf_float
- Upgrade ARM GCC toolchain 12.2.rel1 -> 15.2.rel1 in Dockerfile
- Update LittleFS submodule to v2.11.3

+++ fw/application/src/app/chameleon/scene/chameleon_scene_menu_card_slot_num_select.c
@@ -19,7 +19,7 @@ void chameleon_scene_menu_card_slot_num_select_on_event(mui_list_view_event_t ev
     app_chameleon_t *app = p_list_view->user_data;
     switch (p_item->icon) {
     case ICON_DATA: {
-        tag_helper_set_slot_num((uint8_t)p_item->user_data);
+        tag_helper_set_slot_num((uint8_t)(uintptr_t)p_item->user_data);
         mui_scene_dispatcher_previous_scene(app->p_scene_dispatcher);
     } break;
